backend/pyproject.toml
[tool.poetry]
name = "bot template"
version = "0.1.0"
description = "My template for creating bots with Django+DRF+Aiogram+AioHttp"
authors = ["https://github.com/itisnotyourenv"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
djangorestframework-simplejwt = "^5.0.0"
django-filter = "^21.1"


[tool.poetry.dev-dependencies]
drf-spectacular = "^0.21.1"
drf-spectacular-sidecar = "^2021.12.13"
flake8 = "^4.0.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
